Split data: color maps
