+++ extension/devtools/app/lib/color_maps.dart
@@ -1,0 +1,31 @@
+part of 'main.dart';
+
+const _effectColors = {
+  'UI': OrefPalette.teal,
+  'Network': OrefPalette.indigo,
+  'Persist': OrefPalette.coral,
+  'Analytics': OrefPalette.pink,
+  'Effect': OrefPalette.teal,
+};
+
+const _collectionOpColors = {
+  'Add': OrefPalette.lime,
+  'Remove': OrefPalette.coral,
+  'Replace': OrefPalette.indigo,
+  'Clear': OrefPalette.pink,
+  'Resize': OrefPalette.indigo,
+};
+
+const _deltaStyles = {
+  'add': OrefPalette.lime,
+  'remove': OrefPalette.coral,
+  'update': OrefPalette.indigo,
+};
+
+const _timelineColors = {
+  'signal': OrefPalette.teal,
+  'computed': OrefPalette.indigo,
+  'effect': OrefPalette.pink,
+  'collection': OrefPalette.coral,
+  'batch': OrefPalette.lime,
+};
